pyproject.toml
[project]
name = "genai-chat-bot-backend"
version = "0.1.0"
description = "FastAPI backend for Agentic AI chatbot with LangGraph and ChromaDB"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.6",
  "uvicorn[standard]==0.32.1",
  "pydantic==2.10.3",
  "langchain==0.3.9",
  "langgraph==0.2.56",
  "langchain-groq==0.2.1",
  "langchain-openai==0.2.9",
  "langchain-community==0.3.9",
  "langsmith==0.1.147",
  "chromadb==0.5.20",
  "sentence-transformers==3.3.1",
  "tavily-python==0.5.0",
  "python-dotenv==1.0.1",
  "pytest==8.3.4",
  "httpx==0.28.1",
  "numpy==2.2.0",
  "requests==2.32.3"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "black==24.10.0",
  "flake8==7.1.1",
  "mypy==1.13.0",
  "pytest-cov==6.0.0"
]

[tool.uv]
index-url = "https://pypi.org/simple"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.setuptools.package-data]
"*" = ["*.env", "*.toml", "*.md"]